+++ Runtime/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo(
+    "Deucarian.SimultriaViewerConnection.Tests")]
+[assembly: InternalsVisibleTo(
+    "Deucarian.SimultriaViewerConnection.Editor")]